requirements.txt
Flask==3.1.0
flask-restx==1.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
marshmallow-sqlalchemy==1.4.0
flask-marshmallow==1.3.0
marshmallow-enum==1.5.1
python-dotenv==1.0.1
dynaconf==3.2.7
PyJWT==2.10.1
Werkzeug==3.1.3
Jinja2==3.1.4
psycopg2==2.9.10
PyMySQL==1.1.1
cryptography==44.0.0